+++ plugins/usd_imaging/procedural_custom_adapter.h
@@ -32,6 +32,22 @@ class ArnoldProceduralCustomAdapter : public UsdImagingGprimAdapter {
 public:
     using BaseAdapter = UsdImagingGprimAdapter;
 
+    ArnoldProceduralCustomAdapter()
+        : UsdImagingGprimAdapter()
+    {}
+#ifdef ENABLE_SCENE_INDEX
+    //
+    // Scene index support
+    //
+
+    TfTokenVector GetImagingSubprims(UsdPrim const& prim) override;
+
+    TfToken GetImagingSubprimType(UsdPrim const& prim, TfToken const& subprim) override;
+
+    HdContainerDataSourceHandle GetImagingSubprimData(
+        UsdPrim const& prim, TfToken const& subprim, const UsdImagingDataSourceStageGlobals& stageGlobals) override;
+#endif
+
     /// Populate primitives in the usd imaging index proxy.
     ///
     /// @param prim USD Primitive of the ArnoldProceduralCustom.